pyproject.toml
[project]
name = "multimodal-va-pred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "torch>=2.7.0",
    "torch-geometric>=2.6.1",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
